(PC-17790) fix(types): move SearchListProps inside search types
