+++ src/features/search/components/SearchList/SearchList.tsx
@@ -4,7 +4,7 @@ import styled, { useTheme } from 'styled-components/native'
 
 import { NoSearchResult } from 'features/search/components/NoSearchResults/NoSearchResult'
 import { SearchListHeader } from 'features/search/components/SearchListHeader/SearchListHeader'
-import { SearchListProps } from 'features/search/components/SearchResults/type'
+import { SearchListProps } from 'features/search/types'
 import { SearchHit } from 'libs/algolia'
 import { getSpacing } from 'ui/theme'
