--- src/features/search/components/SearchResults/type.ts
@@ -1,14 +1,0 @@
-import { SearchHit } from 'libs/algolia'
-
-export interface SearchListProps {
-  nbHits: number
-  hits: SearchHit[]
-  renderItem: ({ item, index }: { item: SearchHit; index: number }) => JSX.Element
-  autoScrollEnabled: boolean
-  refreshing: boolean
-  onRefresh: (() => void) | null | undefined
-  isFetchingNextPage: boolean
-  onEndReached: () => void
-  onScroll?: () => void
-  onPress?: () => void
-}
