+++ src/features/search/types.ts
@@ -14,6 +14,7 @@ import {
   MappingTree,
 } from 'features/search/helpers/categoriesHelpers/mapping-tree'
 import { Venue } from 'features/venue/types'
+import { SearchHit } from 'libs/algolia'
 import { SuggestedPlace } from 'libs/place'
 import { Range } from 'libs/typesUtils/typeHelpers'
 
@@ -98,3 +99,16 @@ export type CategoriesViewData =
   | OfferGenreType
 
 export type MappedData = MappingTree | MappedNativeCategories | MappedGenreTypes
+
+export interface SearchListProps {
+  nbHits: number
+  hits: SearchHit[]
+  renderItem: ({ item, index }: { item: SearchHit; index: number }) => JSX.Element
+  autoScrollEnabled: boolean
+  refreshing: boolean
+  onRefresh: (() => void) | null | undefined
+  isFetchingNextPage: boolean
+  onEndReached: () => void
+  onScroll?: () => void
+  onPress?: () => void
+}
